% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{read_text}
\alias{read_text}
\title{Read text files}
\usage{
read_text(x, skip_blank = FALSE)
}
\arguments{
\item{x}{A .txt file to process}

\item{skip_blank}{logical: should blank lines be skipped? Defaults to FALSE}
}
\value{
A character vector
}
\description{
Import .txt files in a standard way
}
