\name{get_file_names}
\alias{get_file_names}
\title{Get names of files within a nested list of folders}
\description{This is a pretty basic implementation, but should work to 3 levels of nestedness.}
\usage{
get_file_names(path)
}
\arguments{
  \item{path}{A path within which to locate files.}
}
\value{
A character vector listing all files and folders
}