\name{read_text}
\alias{read_text}
\title{Read text files}
\description{Import .txt files in a standard way}
\usage{
read_text(x, skip_blank = FALSE)
}
\arguments{
  \item{x}{A .txt file to process}
  \item{skip_blank}{logical: should blank lines be skipped? Defaults to FALSE}
}
\value{
A character vector
}