% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocr_functions.R
\name{run_tesseract}
\alias{run_tesseract}
\title{Run Tesseract}
\usage{
run_tesseract(file, file_out, args = "--psm 1")
}
\arguments{
\item{file}{An image}

\item{file_out}{Where should the .txt file be saved?}

\item{args}{Additional arguments passed to tesseract}
}
\value{
A .txt file.
}
\description{
Process and image using tesseract via a call to 'system'. This requires that
tesseract is already installed on your system.
}
